wlan: Fix MACROS for 32bit to 64bit conversion
This  will fix the MACRO GET_IE_LEN_IN_BSS, GET_IE_LEN_IN_BSS_DESC,
and GET_FIELD_OFFSET. Also removing redundant macro definition
from files to have only definition through out the code.

CRs-fixed: 582104

+++ CORE/HDD/src/wlan_hdd_cfg80211.c
@@ -100,7 +100,7 @@
 #define MAX_SCAN_SSID 9
 #define MAX_PENDING_LOG 5
 #define GET_IE_LEN_IN_BSS_DESC(lenInBss) ( lenInBss + sizeof(lenInBss) - \
-        ((int) OFFSET_OF( tSirBssDescription, ieFields)))
+        ((uintptr_t)OFFSET_OF( tSirBssDescription, ieFields)))
 
 #define HDD2GHZCHAN(freq, chan, flag)   {     \
     .band =  IEEE80211_BAND_2GHZ, \
